Task1/requirements.txt
streamlit
Flask  # For werkzeug.utils (Flask dependency)
faiss-cpu  # CPU version of FAISS, adjust to faiss-gpu if using GPU
numpy
nltk
pymupdf  # Alias for fitz (PyMuPDF)
sentence-transformers
transformers

